pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "paramiko>=3.5.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.4",
    "pysftp>=0.2.9",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]
